+++ app.trading.algoritmico.api/src/AppTradingAlgoritmico.Infrastructure/Persistence/Migrations/20260413010225_SimplifyStatusAndAddRunningDate.Designer.cs
@@ -4,16 +4,19 @@ using AppTradingAlgoritmico.Infrastructure.Persistence;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 #nullable disable
 
-namespace AppTradingAlgoritmico.Infrastructure.Migrations
+namespace AppTradingAlgoritmico.Infrastructure.Persistence.Migrations
 {
     [DbContext(typeof(AppDbContext))]
-    partial class AppDbContextModelSnapshot : ModelSnapshot
+    [Migration("20260413010225_SimplifyStatusAndAddRunningDate")]
+    partial class SimplifyStatusAndAddRunningDate
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder